feat : 골드 획득 구현
- 챕터 및 웨이브 클리어시 골드 획득 및 저장 구현

+++ Heroes_vs_Hordes/Assets/Scripts/Utils/Define.cs
@@ -118,7 +118,7 @@ public class Define
 
     #region IngameManager
     public const float INCREASE_HERO_EXP_VALUE = 1;
-    public const float INCREASE_GOLD_VALUE = 1;
+    public const int INCREASE_GOLD_VALUE = 1;
     public const int INDEX_NORMAL_BATTLE_WAVE = 0;
     public const int INDEX_GOLD_RUSH_WAVE = 1;
     public const int INDEX_BOSS_BATTLE_WAVE = 2;
